Time: 2 ms (57.38%), Space: 8 MB (41.48%) - LeetHub

+++ 0278-first-bad-version/0278-first-bad-version.cpp
@@ -8,14 +8,13 @@ public:
         int h = n;
         int ans = n;
         while(l<=h){
-            long long  mid = (l+ (h-l)/2);
+            int mid = l + (h-l)/2;
             if(isBadVersion(mid)){
                 ans = mid;
                 h = mid-1;
             }else{
                 l = mid+1;
             }
-
         }
         return ans;
     }
